==> notebooks/1_cleaning.ipynb <==
In this notebook, we are going to look at our raw dataset `banks.xlsx` for cleaning purposes. We will be using the `pandas` library to help us with this task.

# import libraries
import numpy as np
import pandas as pd

# load the data
df = pd.read_excel("../data/raw/banks_raw.xlsx")

df.columns

We need to change some columns names, look at the data types ...

df.columns = ['bank_adress', 'bank_category', 'bank_id', 'bank_city', 'bank_latitude', 'bank_longitude', 'bank_phone', 'bank_rank', 'bank_reviews_count', 'bank_title', 'bank_score', 'review_id', 'review_date', 'review_stars', 'review_text', 'review_text_translated', 'reviewer_id', 'reviewer_num_reviews', 'reviewer_name']

df.head()

# drop duplicates reviews
df = df.drop_duplicates()

df['bank_city'].unique()

We notice that there are some cities that are written in different ways, we need to fix that.

df.loc[df['bank_city'] == 'Rabat Hassan', 'bank_city'] = 'Rabat'
df.loc[df['bank_city'] == 'Laayoune', 'bank_city'] = 'Laâyoune'
df['bank_city'].unique()

Cities are now clean.

df['bank_category'].unique()

We see that here, we got some categories that are not really banks, we need to get rid of them.

df = df.loc[df['bank_category'].isin(['Banque', "Banque d'épargne", 'Institution financière', 'Établissement de crédit', np.nan])]
df['bank_category'].unique()

df['bank_title'].unique()

Here, we see that there are many banks with the same name, but written in different ways. We need to fix that. We are going to use the fuzzywuzzy library to help us with that.  
Before that ler's remove the classes that do not refer to banks.

# import the function from utils/functions.py
import sys
sys.path.append('../utils')
import functions

# define a list of bank labels
labels = df['bank_title'].unique()

# create a dictionnary that maps each bank label to its corresponding bank
bank_dict = functions.map_labels_to_banks(labels)

# create a new column that contains the bank name
df['bank_title'] = df['bank_title'].apply(lambda x: bank_dict[x])

# drop the rows that contain unknown banks
df = df.loc[df['bank_title'] != 'unknown']

Next, we look at the other categories related to the bank.

df['bank_phone'].unique()

df['bank_rank'].unique()

Now, all the fields that are related to the bank are tidy.  
We need to do the same for categories related to reviews and reviewers.  
For reviews text, we are only interested in the text itself, we need a column text which holds the text of the review if it's in French, otherwise, we use the translated text, which is in French.

# create a text column which takes text_translated if it exists, otherwise it takes text
df['review_text'] = df['review_text_translated'].fillna(df['review_text'])

# changing reviewer names to unknown if they are nan
df['reviewer_name'] = df['reviewer_name'].fillna('anonymous')

Now we're all set, we can drop the other columns that we don't need.

df.drop(['review_text_translated', 'review_id', 'reviewer_id', 'reviewer_num_reviews'], axis=1, inplace=True)

df['review_id'] = df.index

df.dtypes

Let's work on the date column.

df.loc[df['review_date'].isna()]

# drop values with no date
df = df.loc[df['review_date'].notna()]

df['review_date'].isna().sum()

No missing values, that's good.

# transform the review_date column to datetime
df['review_date'] = pd.to_datetime(df['review_date'], format='%Y-%m-%d', utc=False).dt.date

df.head()

# reorder the columns
df = df[['bank_id', 'bank_title', 'bank_category', 'bank_adress', 'bank_city', 'bank_phone', 'bank_rank', 'bank_reviews_count', 'bank_score', 'bank_latitude', 'bank_longitude', 'review_id', 'review_date', 'review_stars', 'review_text', 'reviewer_name']]

We need to do one more thing in order to make our data warehouse building easier, it's to separate this dataset into 3 datasets, one for banks, one for reviews and one for reviewers.

banks_df = df[['bank_id', 'bank_title', 'bank_category', 'bank_adress', 'bank_city', 'bank_phone', 'bank_rank', 'bank_reviews_count', 'bank_score', 'bank_latitude', 'bank_longitude']]
reviews_df = df[['review_id', 'bank_id', 'review_date', 'review_stars', 'review_text', 'reviewer_name']]

banks_df = banks_df.drop_duplicates()

# create a directory for the processed data (added in .gitignore)
!mkdir ../data/processed

banks_df.to_excel("../data/processed/banks.xlsx", index=False)
reviews_df.to_excel("../data/processed/reviews.xlsx", index=False)
