==> notebooks/2_sentiment-analysis.ipynb <==
In this notebook we are going to load our reviews dataset, apply a bert based model on the reviews texts in order to get the sentiment of each review and then we are going to save the results.

import numpy as np
import pandas as pd

# load the dataset
reviews = pd.read_excel('../data/processed/reviews.xlsx')
reviews.head()

from transformers import pipeline

sentiment_pipeline = pipeline('sentiment-analysis', model='nlptown/bert-base-multilingual-uncased-sentiment')

The approach we'll take here is the following:

- We know that for a review to be published, it must have either a number of stars or a text comment.
- The model for our sentiment analysis gives us an equivalent number of stars for the comment that we give to it
- We are presented against the following cases
  - If we have a text with stars, we are going to take the average between the stars the model give on the text and the stars the user enters
  - If we only have a text, we are only going to use the stars we get from it through the model
  - If we only have stars, we are only going to use them for the sentiment analysis
- When the final stars score is 4 or more, it's positive, if it's 2 or less, it's negative, otherwise it's a neutral review.
- For the missing values of the text, we are going to fill them with the value 'no text'
- For the missing values of stars, we are going to fill them with the value of 0

# prepare the dataset for sentiment analysis
reviews['review_stars'].isna().sum()

# fill missing values
reviews['review_stars'].fillna(0, inplace=True)
reviews['review_text'].fillna('no text', inplace=True)

# define a function for performing the model on the reviews texts
def analyse_text(text):
  # the case of inexistant comment
  if text == 'no text':
    return 0

  # take the first 512 chars from the comment because it's the maximum size the model can handle
  text = text[:512]
  
  # apply model pipeline on text
  result = sentiment_pipeline(text)
  return int(result[0]['label'][0])

# testing the model
analyse_text("j'adore ce produit")

# create stars_from_comment column
reviews['stars_from_comment'] = reviews['review_text'].apply(analyse_text)

reviews.head()

def get_review_score(row):
  star1 = row['review_stars']
  star2 = row['stars_from_comment']
  if star1 == 0:
    return star2
  if star2 == 0:
    return star1
  return (star1 + star2) / 2

# create review_score column
reviews['review_score'] = reviews.apply(get_review_score, axis=1)

reviews.head()

def get_sentiment(score):
  if score <= 2:
    return 'NEGATIVE'
  elif score >= 4:
    return 'POSITIVE'
  return 'NEUTRAL'

# create review_sentiment column
reviews['review_sentiment'] = reviews['review_score'].apply(get_sentiment)

reviews.head()

What is left for us to do is to delete the temporary columns we created, and then save the new dataframe as a csv file.

# drop temporary columns
reviews.drop(['stars_from_comment', 'review_score'], axis=1, inplace=True)

# save the new dataframe
reviews.to_excel("../data/processed/reviews_cleaned.xlsx", index=False)
